chore(validation): move validator file to utilities
- move the validator.py file to the utilities folder
- remove the add_office.py file
- refactor to use validate_empty_fields instead of verify_attributes

[Finishes #163722806]
