+++ api/floor/schema.py
@@ -6,7 +6,7 @@ from helpers.auth.authentication import Auth
 from helpers.auth.admin_roles import admin_roles
 from utilities.validations import validate_empty_fields
 from utilities.utility import update_entity_fields
-from helpers.auth.validator import ErrorHandler
+from utilities.validator import ErrorHandler
 from helpers.pagination.paginate import Paginate, validate_page
 from api.block.models import Block
 from api.floor.models import Floor as FloorModel
